requirements.txt
numpy >= 1.1.0
opencv-python >= 4.0.0
pillow >= 8.0.0
requests >= 2.0.0
